+++ R/functions__lazy.R
@@ -11,7 +11,8 @@
 #' to "all-columns" and is an expression constructor
 #'
 #' @examples
-#' pl$DataFrame(list(all = c(TRUE, TRUE), some = c(TRUE, FALSE)))$select(pl$all()$all())
+#' pl$DataFrame(all = c(TRUE, TRUE), some = c(TRUE, FALSE))$
+#'   select(pl$all()$all())
 pl_all = function(name = NULL) {
   if (is.null(name)) {
     return(.pr$Expr$col("*"))
@@ -62,7 +63,7 @@ pl_all = function(name = NULL) {
 #' df$select(pl$col(pl$dtypes$Float64))
 #'
 #' # ... or an R list of DataTypes, select any column of any such DataType
-#' df$select(pl$col(list(pl$dtypes$Float64, pl$dtypes$Utf8)))
+#' df$select(pl$col(list(pl$dtypes$Float64, pl$dtypes$String)))
 #'
 #' # from Series of names
 #' df$select(pl$col(pl$Series(c("bar", "foobar"))))
@@ -79,8 +80,6 @@ pl_col = function(name = "", ...) {
 #' @examples
 #' pl$lit(1:5)$cumulative_eval(pl$element()$first() - pl$element()$last()**2)$to_r()
 pl_element = function() pl$col("")
-
-# TODO move all lazy functions to a new keyword lazy functions
 
 #' pl$count
 #' @description Count the number of values in this column/context.
@@ -691,13 +690,13 @@ pl_concat_list = function(exprs) {
 #' # wrap two columns in a struct and provide a schema to set all or some DataTypes by name
 #' e1 = pl$struct(
 #'   pl$col(c("int", "str")),
-#'   schema = list(int = pl$Int64, str = pl$Utf8)
+#'   schema = list(int = pl$Int64, str = pl$String)
 #' )$alias("my_struct")
 #' # same result as e.g. wrapping the columns in a struct and casting afterwards
 #' e2 = pl$struct(
 #'   list(pl$col("int"), pl$col("str"))
 #' )$cast(
-#'   pl$Struct(int = pl$Int64, str = pl$Utf8)
+#'   pl$Struct(int = pl$Int64, str = pl$String)
 #' )$alias("my_struct")
 #'
 #' df = pl$DataFrame(
@@ -741,7 +740,7 @@ pl_struct = function(
 #'
 #' @param ... Columns to concatenate into a single string column. Accepts
 #' expressions. Strings are parsed as column names, other non-expression inputs
-#' are parsed as literals. Non-Utf8 columns are cast to Utf8.
+#' are parsed as literals. Non-String columns are cast to String
 #' @param separator String that will be used to separate the values of each
 #' column.
 #' @return Expr
@@ -841,11 +840,9 @@ pl_rolling_corr = function(a, b, window_size, min_periods = NULL, ddof = 1) {
 
 #' Accumulate over multiple columns horizontally with an R function
 #'
-#' @description `pl$fold()` and `pl$reduce()` allows one to do rowwise operations.
-#' The only difference between them is that `pl$fold()` has an additional argument
-#' (`acc`) that contains the value that will be initialized when the fold starts.
-#'
-#' @name pl_fold_reduce
+#' This allows one to do rowwise operations, starting with an initial value
+#' (`acc`). See [`pl$reduce()`][pl_reduce] to do rowwise operations without this initial
+#' value.
 #'
 #' @param acc an Expr or Into<Expr> of the initial accumulator.
 #' @param lambda R function which takes two polars Series as input and return one.
@@ -856,26 +853,35 @@ pl_rolling_corr = function(a, b, window_size, min_periods = NULL, ddof = 1) {
 #' @examples
 #' df = pl$DataFrame(mtcars)
 #'
-#' # Make the row-wise sum of all columns with fold, reduce and vectorized "+"
+#' # Make the row-wise sum of all columns
 #' df$with_columns(
-#'   pl$reduce(
-#'     lambda = \(acc, x) acc + x,
-#'     exprs = pl$col("mpg", "drat")
-#'   )$alias("mpg_drat_sum_reduced"),
 #'   pl$fold(
 #'     acc = pl$lit(0),
 #'     lambda = \(acc, x) acc + x,
-#'     exprs = pl$col("mpg", "drat")
-#'   )$alias("mpg_drat_sum_folded"),
-#'   (pl$col("mpg") + pl$col("drat"))$alias("mpg_drat_vector_sum")
+#'     exprs = pl$col("*")
+#'   )$alias("mpg_drat_sum_folded")
 #' )
 pl_fold = function(acc, lambda, exprs) {
   fold(acc, lambda, exprs) |>
     unwrap("in pl$fold():")
 }
 
-#' @rdname pl_fold_reduce
-#' @name pl_fold_reduce_part2
+#' @inherit pl_fold title params return
+#'
+#' @description
+#' This allows one to do rowwise operations. See [`pl$fold()`][pl_fold] to do rowwise
+#' operations with an initial value.
+#'
+#' @examples
+#' df = pl$DataFrame(mtcars)
+#'
+#' # Make the row-wise sum of all columns
+#' df$with_columns(
+#'   pl$reduce(
+#'     lambda = \(acc, x) acc + x,
+#'     exprs = pl$col("*")
+#'   )$alias("mpg_drat_sum_reduced")
+#' )
 pl_reduce = function(lambda, exprs) {
   reduce(lambda, exprs) |>
     unwrap("in pl$reduce():")