+++ src/rust/src/lazy/dsl.rs
@@ -19,7 +19,7 @@ use extendr_api::{extendr, prelude::*, rprintln, Deref, DerefMut, Rinternals};
 use pl::PolarsError as pl_error;
 use pl::{
     BinaryNameSpaceImpl, Duration, DurationMethods, IntoSeries, RollingGroupOptions,
-    TemporalMethods, Utf8NameSpaceImpl,
+    StringNameSpaceImpl, TemporalMethods,
 };
 use polars::lazy::dsl;
 use polars::prelude as pl;
@@ -133,7 +133,7 @@ impl RPolarsExpr {
             }
             (Rtype::Strings, 1) => {
                 if robj.is_na() {
-                    Ok(dsl::lit(pl::NULL).cast(pl::DataType::Utf8))
+                    Ok(dsl::lit(pl::NULL).cast(pl::DataType::String))
                 } else {
                     Ok(dsl::lit(robj.as_str().unwrap()))
                 }
@@ -299,7 +299,11 @@ impl RPolarsExpr {
     }
 
     pub fn gather(&self, idx: Robj) -> RResult<Self> {
-        Ok(self.clone().0.gather(robj_to!(PLExpr, idx)?).into())
+        Ok(self
+            .clone()
+            .0
+            .gather(robj_to!(PLExpr, idx)?.cast(pl::DataType::Int64))
+            .into())
     }
 
     pub fn sort_by(&self, by: Robj, descending: Robj) -> RResult<RPolarsExpr> {
@@ -919,7 +923,7 @@ impl RPolarsExpr {
                     move |s| {
                         //swap owned inline string to str as only supported and if swapped here life time is long enough
                         let av = match &av {
-                            pl::AnyValue::Utf8Owned(x) => pl::AnyValue::Utf8(x.as_str()),
+                            pl::AnyValue::StringOwned(x) => pl::AnyValue::String(x.as_str()),
                             x => x.clone(),
                         };
                         s.extend_constant(av, n).map(Some)
@@ -1007,6 +1011,14 @@ impl RPolarsExpr {
                 robj_to!(Option, PLPolarsDataType, return_dtype)?,
             )
             .into())
+    }
+
+    pub fn rle(&self) -> RResult<Self> {
+        Ok(self.0.clone().rle().into())
+    }
+
+    pub fn rle_id(&self) -> RResult<Self> {
+        Ok(self.0.clone().rle_id().into())
     }
 
     //arr/list methods
@@ -1824,7 +1836,7 @@ impl RPolarsExpr {
     pub fn str_len_bytes(&self) -> Self {
         use pl::*;
         let function = |s: pl::Series| {
-            let ca = s.utf8()?;
+            let ca = s.str()?;
             Ok(Some(ca.str_len_bytes().into_series()))
         };
         self.clone()
@@ -1836,7 +1848,7 @@ impl RPolarsExpr {
 
     pub fn str_len_chars(&self) -> Self {
         let function = |s: pl::Series| {
-            let ca = s.utf8()?;
+            let ca = s.str()?;
             Ok(Some(ca.str_len_chars().into_series()))
         };
         self.clone()
@@ -1941,7 +1953,7 @@ impl RPolarsExpr {
             use pl::*;
             let pat: String = robj_to!(String, pat, "in str$json_path_match: {}")?;
             let function = move |s: Series| {
-                let ca = s.utf8()?;
+                let ca = s.str()?;
                 match ca.json_path_match(&pat) {
                     Ok(ca) => Ok(Some(ca.into_series())),
                     Err(e) => Err(pl::PolarsError::ComputeError(format!("{e:?}").into())),
@@ -1950,7 +1962,7 @@ impl RPolarsExpr {
             Ok(RPolarsExpr(
                 self.0
                     .clone()
-                    .map(function, pl::GetOutput::from_type(pl::DataType::Utf8))
+                    .map(function, pl::GetOutput::from_type(pl::DataType::String))
                     .with_fmt("str.json_path_match"),
             ))
         }();
@@ -1968,65 +1980,38 @@ impl RPolarsExpr {
             .into())
     }
 
-    pub fn str_hex_encode(&self) -> Self {
-        use pl::*;
-        self.clone()
-            .0
-            .map(
-                move |s| s.utf8().map(|s| Some(s.hex_encode().into_series())),
-                pl::GetOutput::same_type(),
-            )
-            .with_fmt("str.hex_encode")
-            .into()
+    pub fn str_hex_encode(&self) -> RResult<Self> {
+        Ok(self.0.clone().str().hex_encode().into())
     }
 
-    pub fn str_hex_decode(&self, strict: bool) -> Self {
-        use pl::*;
-        self.clone()
+    pub fn str_hex_decode(&self, strict: Robj) -> RResult<Self> {
+        Ok(self
             .0
-            .map(
-                move |s| s.utf8()?.hex_decode(strict).map(|s| Some(s.into_series())),
-                pl::GetOutput::same_type(),
-            )
-            .with_fmt("str.hex_decode")
-            .into()
+            .clone()
+            .str()
+            .hex_decode(robj_to!(bool, strict)?)
+            .into())
     }
-    pub fn str_base64_encode(&self) -> Self {
-        use pl::*;
-        self.clone()
-            .0
-            .map(
-                move |s| s.utf8().map(|s| Some(s.base64_encode().into_series())),
-                pl::GetOutput::same_type(),
-            )
-            .with_fmt("str.base64_encode")
-            .into()
+    pub fn str_base64_encode(&self) -> RResult<Self> {
+        Ok(self.0.clone().str().base64_encode().into())
     }
 
-    pub fn str_base64_decode(&self, strict: bool) -> Self {
-        use pl::*;
-        self.clone()
+    pub fn str_base64_decode(&self, strict: Robj) -> RResult<Self> {
+        Ok(self
             .0
-            .map(
-                move |s| {
-                    s.utf8()?
-                        .base64_decode(strict)
-                        .map(|s| Some(s.into_series()))
-                },
-                pl::GetOutput::same_type(),
-            )
-            .with_fmt("str.base64_decode")
-            .into()
+            .clone()
+            .str()
+            .base64_decode(robj_to!(bool, strict)?)
+            .into())
     }
 
-    pub fn str_extract(&self, pattern: Robj, group_index: Robj) -> List {
-        let res = || -> Result<RPolarsExpr, String> {
-            let pat = robj_to!(String, pattern)?;
-            let gi = robj_to!(usize, group_index)?;
-            Ok(self.0.clone().str().extract(pat.as_str(), gi).into())
-        }()
-        .map_err(|err| format!("in str$extract: {}", err));
-        r_result_list(res)
+    pub fn str_extract(&self, pattern: Robj, group_index: Robj) -> RResult<Self> {
+        Ok(self
+            .0
+            .clone()
+            .str()
+            .extract(robj_to!(str, pattern)?, robj_to!(usize, group_index)?)
+            .into())
     }
 
     pub fn str_extract_all(&self, pattern: &RPolarsExpr) -> Self {
@@ -2138,6 +2123,41 @@ impl RPolarsExpr {
             .into())
     }
 
+    pub fn str_reverse(&self) -> RResult<Self> {
+        Ok(self.0.clone().str().reverse().into())
+    }
+
+    pub fn str_contains_any(&self, patterns: Robj, ascii_case_insensitive: Robj) -> RResult<Self> {
+        Ok(self
+            .0
+            .clone()
+            .str()
+            .contains_any(
+                robj_to!(PLExpr, patterns)?,
+                robj_to!(bool, ascii_case_insensitive)?,
+            )
+            .into())
+    }
+
+    pub fn str_replace_many(
+        &self,
+        patterns: Robj,
+        replace_with: Robj,
+        ascii_case_insensitive: Robj,
+    ) -> RResult<Self> {
+        Ok(self
+            .0
+            .clone()
+            .str()
+            .replace_many(
+                robj_to!(PLExpr, patterns)?,
+                robj_to!(PLExpr, replace_with)?,
+                robj_to!(bool, ascii_case_insensitive)?,
+            )
+            .into())
+    }
+
+    //binary methods
     pub fn bin_contains(&self, lit: Robj) -> RResult<Self> {
         Ok(self
             .0