+++ tests/testthat/test-dataframe.R
@@ -144,13 +144,13 @@ test_that("get set properties", {
 test_that("DataFrame, custom schema", {
   df = pl$DataFrame(
     iris,
-    schema = list(Sepal.Length = pl$Float32, Species = pl$Utf8)
+    schema = list(Sepal.Length = pl$Float32, Species = pl$String)
   )
   # dtypes from object are as expected
   expect_true(
     all(mapply(
       df$dtypes,
-      pl$dtypes[c("Float32", rep("Float64", 3), "Utf8")],
+      pl$dtypes[c("Float32", rep("Float64", 3), "String")],
       FUN = "=="
     ))
   )
@@ -1217,7 +1217,7 @@ test_that("transpose", {
   expect_identical(
     pl$DataFrame(iris)$
       with_columns(pl$col("Species")$
-      cast(pl$Utf8))$
+      cast(pl$String))$
       transpose(FALSE)$
       to_data_frame(),
     df_expected
@@ -1236,4 +1236,45 @@ test_that("drop_all_in_place", {
   expect_identical(df_copy$shape, c(0, 0))
   expect_identical(df_clone$shape, c(32, 11))
   expect_identical(s$len(), 32)
+})
+
+test_that("rolling for DataFrame: basic example", {
+  # this is just to ensure that the DataFrame method calls the lazy method
+  # under the hood. See the tests on lazy for more.
+  df = pl$DataFrame(
+    dt = c(
+      "2020-01-01", "2020-01-01", "2020-01-01",
+      "2020-01-02", "2020-01-03", "2020-01-08"
+    ),
+    a = c(3, 7, 5, 9, 2, 1)
+  )$with_columns(
+    pl$col("dt")$str$strptime(pl$Date, format = NULL)$set_sorted()
+  )
+
+  actual = df$rolling(index_column = "dt", period = "2d")$agg(
+    pl$sum("a")$alias("sum_a"),
+    pl$min("a")$alias("min_a"),
+    pl$max("a")$alias("max_a")
+  )$to_data_frame()
+
+  expect_equal(
+    actual[, c("sum_a", "min_a", "max_a")],
+    data.frame(
+      sum_a = c(15, 15, 15, 24, 11, 1),
+      min_a = c(3, 3, 3, 3, 2, 1),
+      max_a = c(7, 7, 7, 9, 9, 1)
+    )
+  )
+})
+
+test_that("rolling for DataFrame: can be ungrouped", {
+  df = pl$DataFrame(
+    index = c(1:5, 6.0),
+    a = c(3, 7, 5, 9, 2, 1)
+  )$with_columns(pl$col("index")$set_sorted())
+
+  actual = df$rolling(index_column = "dt", period = "2i")$
+    ungroup()$
+    to_data_frame()
+  expect_equal(actual, df$to_data_frame())
 })